added details page and updated all_products function in views.py

+++ store/models.py
@@ -3,28 +3,29 @@ from django.urls import reverse
 from accounts.models import Account
 # Create your models here.
 
-RATING=(
+RATING = (
     (1, '1'),
     (2, '2'),
     (3, '3'),
     (4, '4'),
-    (5, '5'),    
+    (5, '5'),
 )
 
+
 class Review(models.Model):
-    account=models.ForeignKey(Account,on_delete=models.CASCADE)
+    account = models.ForeignKey(Account, on_delete=models.CASCADE)
     rating: models.CharField(choices=RATING, max_length=10)
     subject: models.CharField(max_length=100)
     content: models.TextField(max_length=300)
+
 
 class Product(models.Model):
     name = models.CharField(max_length=255)
     slug = models.SlugField(max_length=255)
     description = models.TextField(blank=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=6, decimal_places=2)
     image = models.ImageField(blank=True, upload_to='images/')
     stock = models.PositiveIntegerField()
-
 
     # required
 
@@ -32,7 +33,7 @@ class Product(models.Model):
     is_active = models.BooleanField(default=True)
     created_date = models.DateTimeField(auto_now_add=True)
     modified_date = models.DateTimeField(auto_now_add=True)
-    review=models.ForeignKey(Review, on_delete=models.CASCADE, null=True)
+    review = models.ForeignKey(Review, on_delete=models.CASCADE, null=True)
 
     class Meta:
         verbose_name_plural = 'Products'
